+++ src/main/java/com/epam/reportportal/service/ReportPortal.java
@@ -25,7 +25,6 @@ import com.epam.reportportal.utils.SslUtils;
 import com.epam.reportportal.utils.files.Utils;
 import com.epam.reportportal.utils.http.ClientUtils;
 import com.epam.reportportal.utils.http.HttpRequestUtils;
-import com.epam.reportportal.utils.properties.ListenerProperty;
 import com.epam.reportportal.utils.properties.PropertiesLoader;
 import com.epam.ta.reportportal.ws.model.launch.StartLaunchRQ;
 import com.epam.ta.reportportal.ws.model.log.SaveLogRQ;
@@ -38,6 +37,7 @@ import okhttp3.OkHttpClient;
 import okhttp3.logging.HttpLoggingInterceptor;
 import org.apache.commons.lang3.BooleanUtils;
 import org.apache.commons.lang3.StringUtils;
+import org.apache.commons.lang3.exception.ExceptionUtils;
 import org.slf4j.Logger;
 import org.slf4j.LoggerFactory;
 import retrofit2.Retrofit;
@@ -51,9 +51,7 @@ import java.io.File;
 import java.io.IOException;
 import java.net.MalformedURLException;
 import java.net.URL;
-import java.security.KeyManagementException;
-import java.security.KeyStoreException;
-import java.security.NoSuchAlgorithmException;
+import java.security.*;
 import java.util.*;
 import java.util.concurrent.ConcurrentHashMap;
 import java.util.concurrent.ExecutorService;
@@ -64,8 +62,8 @@ import java.util.function.Function;
 
 import static com.epam.reportportal.service.LaunchLoggingContext.DEFAULT_LAUNCH_KEY;
 import static com.epam.reportportal.utils.ObjectUtils.clonePojo;
+import static com.epam.reportportal.utils.formatting.ExceptionUtils.getStackTrace;
 import static java.util.Optional.ofNullable;
-import static org.apache.commons.lang3.exception.ExceptionUtils.getStackTrace;
 
 /**
  * Default ReportPortal Reporter implementation. Uses
@@ -313,69 +311,14 @@ public class ReportPortal {
 		});
 	}
 
-	private static void fillSaveLogRQ(final SaveLogRQ rq, final String message, final String level, final Date time, final File file) {
-		rq.setMessage(message);
-		rq.setLevel(level);
-		rq.setLogTime(time);
-
-		try {
-			SaveLogRQ.File f = new SaveLogRQ.File();
-			TypeAwareByteSource typedSource = Utils.getFile(file);
-			f.setContentType(typedSource.getMediaType());
-			f.setContent(typedSource.read());
-
-			f.setName(UUID.randomUUID().toString());
-			rq.setFile(f);
-		} catch (IOException e) {
-			// seems like there is some problem. Do not report a file
-			LOGGER.error("Cannot send file to ReportPortal", e);
-		}
-	}
-
-	/**
-	 * Emits log message if there is any active context attached to the current thread
-	 *
-	 * @param message Log message
-	 * @param level   Log level
-	 * @param time    Log time
-	 * @param file    a file to attach to the log message
-	 * @return true if log has been emitted
-	 */
-	public static boolean emitLog(final String message, final String level, final Date time, final File file) {
-		return emitLog(itemUuid -> {
-			SaveLogRQ rq = new SaveLogRQ();
-			rq.setItemUuid(itemUuid);
-			fillSaveLogRQ(rq, message, level, time, file);
-			return rq;
-		});
-	}
-
-	/**
-	 * Emits log message on Launch level if there is any active context attached to the current thread
-	 *
-	 * @param message Log message
-	 * @param level   Log level
-	 * @param time    Log time
-	 * @param file    a file to attach to the log message
-	 * @return true if log has been emitted
-	 */
-	public static boolean emitLaunchLog(final String message, final String level, final Date time, final File file) {
-		return emitLaunchLog(launchUuid -> {
-			SaveLogRQ rq = new SaveLogRQ();
-			rq.setLaunchUuid(launchUuid);
-			fillSaveLogRQ(rq, message, level, time, file);
-			return rq;
-		});
-	}
-
 	/**
 	 * Converts {@link ReportPortalMessage} to {@link SaveLogRQ}.
 	 *
 	 * @param launchUuid a UUID of the launch
-	 * @param itemUuid a UUID of the test item
-	 * @param level message level
-	 * @param time timestamp of the message
-	 * @param message an object to convert
+	 * @param itemUuid   a UUID of the test item
+	 * @param level      message level
+	 * @param time       timestamp of the message
+	 * @param message    an object to convert
 	 * @return converted object
 	 */
 	@Nonnull
@@ -403,11 +346,53 @@ public class ReportPortal {
 	}
 
 	/**
+	 * Emits log message if there is any active context attached to the current thread
+	 *
+	 * @param message Log message
+	 * @param level   Log level
+	 * @param time    Log time
+	 * @param file    a file to attach to the log message
+	 * @return true if log has been emitted
+	 */
+	public static boolean emitLog(final String message, final String level, final Date time, final File file) {
+		return emitLog(itemUuid -> {
+			try {
+				TypeAwareByteSource byteSource = Utils.getFile(file);
+				return toSaveLogRQ(null, itemUuid, level, time, new ReportPortalMessage(byteSource, message));
+			} catch (IOException e) {
+				LOGGER.error("Cannot read file", e);
+				return toSaveLogRQ(null, itemUuid, level, time, new ReportPortalMessage(message));
+			}
+		});
+	}
+
+	/**
+	 * Emits log message on Launch level if there is any active context attached to the current thread
+	 *
+	 * @param message Log message
+	 * @param level   Log level
+	 * @param time    Log time
+	 * @param file    a file to attach to the log message
+	 * @return true if log has been emitted
+	 */
+	public static boolean emitLaunchLog(final String message, final String level, final Date time, final File file) {
+		return emitLaunchLog(launchUuid -> {
+			try {
+				TypeAwareByteSource byteSource = Utils.getFile(file);
+				return toSaveLogRQ(launchUuid, null, level, time, new ReportPortalMessage(byteSource, message));
+			} catch (IOException e) {
+				LOGGER.error("Cannot read file", e);
+				return toSaveLogRQ(launchUuid, null, level, time, new ReportPortalMessage(message));
+			}
+		});
+	}
+
+	/**
 	 * Emit log message to the current test item.
 	 *
 	 * @param message an instance of the message
-	 * @param level message level
-	 * @param time timestamp of the message
+	 * @param level   message level
+	 * @param time    timestamp of the message
 	 * @return true if log has been emitted otherwise false
 	 */
 	public static boolean emitLog(final ReportPortalMessage message, final String level, final Date time) {
@@ -418,8 +403,8 @@ public class ReportPortal {
 	 * Emit log message to the current Launch.
 	 *
 	 * @param message an instance of the message
-	 * @param level message level
-	 * @param time timestamp of the message
+	 * @param level   message level
+	 * @param time    timestamp of the message
 	 * @return true if log has been emitted otherwise false
 	 */
 	public static boolean emitLaunchLog(final ReportPortalMessage message, final String level, final Date time) {
@@ -432,13 +417,18 @@ public class ReportPortal {
 	 * @param cause a {@link Throwable}
 	 */
 	public static void sendStackTraceToRP(final Throwable cause) {
+		ListenerParameters myParameters = ofNullable(Launch.currentLaunch()).map(Launch::getParameters).orElseGet(ListenerParameters::new);
 		ReportPortal.emitLog(itemUuid -> {
 			SaveLogRQ rq = new SaveLogRQ();
 			rq.setItemUuid(itemUuid);
 			rq.setLevel("ERROR");
 			rq.setLogTime(Calendar.getInstance().getTime());
 			if (cause != null) {
-				rq.setMessage(getStackTrace(cause));
+				if (myParameters.isExceptionTruncate()) {
+					rq.setMessage(getStackTrace(cause, new Throwable()));
+				} else {
+					rq.setMessage(ExceptionUtils.getStackTrace(cause));
+				}
 			} else {
 				rq.setMessage("Test has failed without exception");
 			}
@@ -556,37 +546,68 @@ public class ReportPortal {
 
 			String keyStore = parameters.getKeystore();
 			String keyStorePassword = parameters.getKeystorePassword();
+			String trustStore = parameters.getTruststore();
+			String trustStorePassword = parameters.getTruststorePassword();
 
 			OkHttpClient.Builder builder = new OkHttpClient.Builder();
 
-			if (HTTPS.equals(baseUrl.getProtocol()) && keyStore != null) {
-				if (null == keyStorePassword) {
-					String error = "You should provide keystore password parameter [" + ListenerProperty.KEYSTORE_PASSWORD
-							+ "] if you use HTTPS protocol";
-					LOGGER.error(error);
-					throw new InternalReportPortalClientException(error);
+			if (HTTPS.equals(baseUrl.getProtocol()) && (keyStore != null || trustStore != null)) {
+				KeyManager[] keyManagers = null;
+				if (keyStore != null) {
+					KeyStore ks = SslUtils.loadKeyStore(keyStore, keyStorePassword);
+					try {
+						KeyManagerFactory kmf = KeyManagerFactory.getInstance(KeyManagerFactory.getDefaultAlgorithm());
+						kmf.init(ks, ofNullable(keyStorePassword).map(String::toCharArray).orElse(null));
+						keyManagers = kmf.getKeyManagers();
+					} catch (KeyStoreException | NoSuchAlgorithmException | UnrecoverableKeyException e) {
+						String error = "Unable to load key store";
+						LOGGER.error(error, e);
+						throw new InternalReportPortalClientException(error, e);
+					}
+				}
+
+				TrustManager[] trustManagers = null;
+				if (trustStore != null) {
+					KeyStore ts = SslUtils.loadKeyStore(trustStore, trustStorePassword);
+					try {
+						TrustManagerFactory tmf = TrustManagerFactory.getInstance(TrustManagerFactory.getDefaultAlgorithm());
+						tmf.init(ts);
+						trustManagers = tmf.getTrustManagers();
+					} catch (KeyStoreException | NoSuchAlgorithmException e) {
+						String trustStoreError = "Unable to load trust store";
+						LOGGER.error(trustStoreError, e);
+						throw new InternalReportPortalClientException(trustStoreError, e);
+					}
+				}
+
+				if (trustManagers == null) {
+					try {
+						TrustManagerFactory tmf = TrustManagerFactory.getInstance(TrustManagerFactory.getDefaultAlgorithm());
+						tmf.init((KeyStore) null);
+						trustManagers = tmf.getTrustManagers();
+					} catch (NoSuchAlgorithmException | KeyStoreException e) {
+						String trustStoreError = "Unable to load default trust store";
+						LOGGER.error(trustStoreError, e);
+						throw new InternalReportPortalClientException(trustStoreError, e);
+					}
 				}
 
 				try {
-					TrustManagerFactory trustManagerFactory = TrustManagerFactory.getInstance(TrustManagerFactory.getDefaultAlgorithm());
-					trustManagerFactory.init(SslUtils.loadKeyStore(keyStore, keyStorePassword));
-					TrustManager[] trustManagers = trustManagerFactory.getTrustManagers();
-					X509TrustManager trustManager = (X509TrustManager) ofNullable(trustManagers).flatMap(managers -> Arrays.stream(managers)
-									.filter(m -> m instanceof X509TrustManager)
-									.findAny())
-							.orElseThrow(() -> new InternalReportPortalClientException(
-									"Unable to find X509 trust manager, managers:" + Arrays.toString(trustManagers)));
-
 					SSLContext sslContext = SSLContext.getInstance("TLS");
-					sslContext.init(null, new TrustManager[] { trustManager }, null);
-					SSLSocketFactory sslSocketFactory = sslContext.getSocketFactory();
-					builder.sslSocketFactory(sslSocketFactory, trustManager);
-				} catch (KeyStoreException | NoSuchAlgorithmException | KeyManagementException e) {
-					String error = "Unable to load trust store";
+					sslContext.init(keyManagers, trustManagers, new SecureRandom());
+					X509TrustManager trustManager = Arrays.stream(ofNullable(trustManagers).orElse(new TrustManager[] {}))
+							.filter(m -> m instanceof X509TrustManager)
+							.map(m -> (X509TrustManager) m)
+							.findAny()
+							.orElseThrow(() -> new InternalReportPortalClientException("Unable to find X509 trust manager"));
+					builder.sslSocketFactory(sslContext.getSocketFactory(), trustManager);
+				} catch (NoSuchAlgorithmException | KeyManagementException e) {
+					String error = "Unable to initialize SSL context";
 					LOGGER.error(error, e);
 					throw new InternalReportPortalClientException(error, e);
 				}
 			}
+
 			ClientUtils.setupProxy(builder, parameters);
 			builder.addInterceptor(new BearerAuthInterceptor(parameters.getApiKey()));
 			builder.addInterceptor(new PathParamInterceptor("projectName", parameters.getProjectName()));